+++ src/Shipment/resources/database/migrations/2023_03_28_195634_add_carrier_cost_to_shipments_table.php
@@ -1,0 +1,22 @@
+<?php
+
+declare(strict_types=1);
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+
+return new class () extends Migration {
+    public function up()
+    {
+        Schema::table('shipments', function (Blueprint $table) {
+            $table->decimal('carrier_cost', 15, 4)->nullable()->after('configuration');
+        });
+    }
+
+    public function down()
+    {
+        Schema::table('shipments', function (Blueprint $table) {
+            $table->dropColumn('carrier_cost');
+        });
+    }
+};